+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 ## [Unreleased]
 
 -   disableLanguages only in User / Workspace settings (no more in folder settings).It allows to register formatters for every supported language.
+-   Prettier [1.14](https://prettier.io/blog/2018/07/29/1.14.0.html)
 -   New option: tslintIntegration (boolean) use `prettier-tslint` instead of `prettier`
 
 ## [1.5.0]