+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 
 -   Validate the `"prettier"` key in `package.json` using the prettier settings schema
 -   Prettier [1.15](https://prettier.io/blog/2018/11/07/1.15.0.html)
+-   New options: jsxSingleQuote, htmlWhitespaceSensitivity and endOfLine (More info in readme).
 
 ## [1.6.0]